% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adj2dag}
\alias{adj2dag}
\title{Remove edges to satisfy dag constrains on graph}
\usage{
adj2dag(adj_matrix, topE = 150)
}
\arguments{
\item{adj_matrix}{adjacency matrix}

\item{topE}{number of top edges selected. default 150.}
}
\description{
This function does something amazing.
}
