% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkVisualize_layout.R
\name{network_visualize}
\alias{network_visualize}
\title{Visualize network based on adjacency matrix}
\usage{
network_visualize(
  graph,
  gvarType,
  gvarShape,
  truth = NULL,
  mcoord = NULL,
  topE = 50,
  layout.name = "fruchtermanreingold",
  alpha.range = c(0.05, 1),
  edge.weight = T,
  use.alpha = F,
  sym = F,
  label.color = "white",
  seed = 42,
  nodeLabel = T
)
}
\arguments{
\item{graph}{adjacency matrix for graph.}

\item{gvarType}{variable group1. annotated in different colors.}

\item{gvarShape}{variable group2. annotated in different shapes.}

\item{truth}{truth matrix for edge labeling.}

\item{mcoord}{coordinates for plot embedding.}

\item{topE}{number of top edges to be plotted.}
}
\value{
network graph and node coordinates.
}
\description{
network visualization function.
}
